% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compact_train}
\alias{compact_train}
\title{Data used to model the DESPLIM compactness metric}
\format{
A data frame with 547 rows and 9 columns:
\describe{
\item{compact}{Human-assigned compactness score, rescaled to 0 (least compact) to 1 (most compact)}
\item{boyce}{Boyce-Clark index: mean deviation of 16 evenly spaced radials from the interior centre (0-Inf, 0 = most compact)}
\item{box_reock}{Bounding Box Reock: ratio of district area to its minimum bounding rectangle area (0-1)}
\item{hull}{Convex Hull: ratio of district area to its convex hull area (0-1)}
\item{len_wid}{Length-Width: ratio of district width to length via bounding box (0-1, 1 = equal dimensions)}
\item{polsby}{Polsby-Popper: proportional to the ratio of district area to its perimeter squared (0-1)}
\item{skew}{Skew: ratio of maximum inscribed circle area to minimum bounding circle area (0-1)}
\item{sym_x}{X-symmetry: intersection of district with its X-axis reflection, relative to total area (0-1)}
\item{sym_y}{Y-symmetry: intersection of district with its Y-axis reflection, relative to total area (0-1)}
}
}
\source{
\url{https://dataverse.harvard.edu/file.xhtml?fileId=4143644&version=1.1}

\url{https://github.com/aaronrkaufman/compactness/tree/e91519e15fd3ca55f57fc21323bb12ed5a66305b/data}

\url{https://alarm-redist.org/redistmetrics/}
}
\usage{
compact_train
}
\description{
Clean data used to model the DESPLIM compactness metric, based on
the data provided by Kaufman et al. (2021). The metrics have been computed
using the \code{redistmetrics} package.
}
\keyword{datasets}
