% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desplim_angles.R
\name{desplim_angles}
\alias{desplim_angles}
\title{Calculate the minimum and maximum angles between two \code{sf} LINESTRING objects}
\usage{
desplim_angles(input_lines, reference_lines, cast_substring = TRUE)
}
\arguments{
\item{input_lines}{object of class sf of type LINESTRING or MULTILINESTRING.}

\item{reference_lines}{object of class sf of type LINESTRING or
MULTILINESTRING.}

\item{cast_substring}{logical; should the input linestrings be cast to
substring? Default is \code{TRUE}.}
}
\value{
An sf object with the geometries from \code{input_lines} with added
columns containing the values of the minimum and maximum angle.
}
\description{
Function for calculating the minimum and maximum angles
between two intersecting \code{sf} MULTILINESTRING or LINESTRING objects.
}
\details{
If the input and reference lines are already cast to substring,
the function can be sped up, notably by setting \code{cast_substring} to \code{FALSE}
for larger geometries.
}
