% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{desplim_compact_data}
\alias{desplim_compact_data}
\title{DESPLIM compactness data}
\format{
\subsection{\code{desplim_compact_data}}{

A data frame with 558 rows and 11 columns:
\describe{
\item{compact}{District compactness}
\item{boyce}{Boyce Clark ratio}
\item{box_reock}{Box Reock compactness}
\item{hull}{Convex Hull compactness}
\item{len_wid}{Length Width compactness}
\item{polsby}{Polsby Popper compactness}
\item{reock}{Reock compactness}
\item{schwartz}{Schwartzberg compactness}
\item{skew}{Skew compactness}
\item{sym_x}{X symmetry compactness}
\item{sym_y}{Y symmetry compactness}
}
}
}
\source{
\url{https://dataverse.harvard.edu/file.xhtml?fileId=4143644&version=1.1}

\url{https://github.com/aaronrkaufman/compactness/tree/e91519e15fd3ca55f57fc21323bb12ed5a66305b/data}

\url{https://alarm-redist.org/redistmetrics/}
}
\usage{
desplim_compact_data
}
\description{
Clean data used to model the DESPLIM compactness metric, based on
the data provided by Kaufmann et al. (2021). The metrics have been computed
using the \code{redistmetrics} package.
}
\keyword{datasets}
