% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desplim_compactness.R
\name{desplim_compactness}
\alias{desplim_compactness}
\title{Calculate the 'desplim' compactness of an \code{sf} POLYGON object}
\usage{
desplim_compactness(input, keep_metrics = FALSE)
}
\arguments{
\item{input}{object of class \code{sf} of type POLYGON or MULTIPOLYGON.}

\item{keep_metrics}{logical; should all the compactness metrics be kept?
Default is \code{FALSE}.}
}
\value{
An sf object of type POLYGON matching \code{input}, with the compactness
metric(s) added.
}
\description{
Based on the approach described in Kaufmann et al. (2018), this
function computes the compactness of an \code{sf} POLYGON object.
}
\details{
Please see the dedicated vignette for information about the
compactness metric. If \code{input} contains geometries of type MULTIPOLYGON, they
are cast to POLYGON before computing.
}
