% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desplim_split_merge.R
\name{desplim_split_merge}
\alias{desplim_split_merge}
\title{Apply the DESPLIM algorithm to a set of input polygons and lines.}
\usage{
desplim_split_merge(
  input_polygon,
  input_lines,
  input_buildings = NULL,
  line_type_identifier = NULL,
  line_type_hierarchy = NULL,
  parallel = FALSE,
  max_iter = Inf,
  compact_allow_subsequent = 1,
  ...
)
}
\arguments{
\item{input_polygon}{object of class sf of type POLYGON to be merged.}

\item{input_lines}{object of class sf of type LINESTRING or MULTILINESTRING
to be used for splitting.}

\item{input_buildings}{object of class sf of type POLYGON or MULTIPOLYGON.
An optional sf object representing buildings which should be considered when
splitting and merging. Default is \code{NULL}.}

\item{line_type_identifier}{string; the name of the column in \code{input_lines}
which contains the line type. Default is \code{NULL}.}

\item{line_type_hierarchy}{list; hierarchical structure of line types used
for splitting. Default is \code{NULL}. All types can be accessed with \code{"all"}.}

\item{parallel}{logical; whether to run the split and merging steps in
parallel. If \code{TRUE}, uses the \code{future} package, where the number of workers
should be set using the \code{plan} argument. Default is \code{FALSE}.}

\item{max_iter}{integer; maximum number of split-merge iterations to perform
per hierarchy level. The algorithm repeats split-merge until the polygon set
no longer changes (convergence) or \code{max_iter} is reached. Default is \code{Inf}.}

\item{compact_allow_subsequent}{numerical; value of \code{compact_allow} passed to
\code{desplim_merge} from the second iteration onward. A higher value is more
lenient, reducing the risk of re-merging polygons created in earlier
iterations. Only applied when \code{max_iter > 1}. Default is \code{1} (compactness
does not trigger merges in subsequent iterations).}

\item{...}{additional arguments passed to \code{desplim_split}, \code{desplim_merge}
and \code{desplim_connect_border}.}
}
\value{
An sf object of POLYGONs resulting from the merge.
}
\description{
Apply the DESPLIM algorithm to a set of input polygons and lines.
}
\details{
The function applies the DESPLIM algorithm to a set of input
polygons and lines. For each hierarchy level, split-merge is repeated until
convergence (the polygon set no longer changes between iterations) or
\code{max_iter} is reached. Convergence is determined by the number of output
polygons being equal to the previous iteration.
}
\note{
When using \code{parallel = TRUE}, each worker session will print the \code{sf}
startup message on first load. To suppress this, configure your \code{future}
plan with an initializer before calling this function:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{future::plan(future::multisession, workers = n,
  initializer = function() suppressPackageStartupMessages(library(sf)))
}\if{html}{\out{</div>}}
}
