% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kaufman_25}
\alias{kaufman_25}
\title{Example districts from the compactness training data}
\format{
\subsection{\code{kaufman_25}}{

A sf data frame with 25 rows and 2 columns:
\describe{
\item{compact}{District compactness}
\item{geometry}{District shape}
}
}
}
\source{
\url{https://dataverse.harvard.edu/file.xhtml?fileId=4143644&version=1.1}
}
\usage{
kaufman_25
}
\description{
Example districts with compactness score from the data provided by
Kaufmann et al. (2021).
}
\keyword{datasets}
