% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compact_train}
\alias{compact_train}
\title{Data used to model the DESPLIM compactness metric}
\format{
\subsection{\code{compact_train}}{

A data frame with 547 rows and 11 columns:
\describe{
\item{compact}{Kaufman compactness}
\item{boyce}{Boyce Clark ratio}
\item{box_reock}{Box Reock compactness}
\item{hull}{Convex Hull compactness}
\item{len_wid}{Length-Width compactness}
\item{polsby}{Polsby Popper compactness}
\item{reock}{Reock compactness}
\item{schwartz}{Schwartzberg compactness}
\item{skew}{Skew compactness}
\item{sym_x}{X symmetry}
\item{sym_y}{Y symmetry}
}
}
}
\source{
\url{https://dataverse.harvard.edu/file.xhtml?fileId=4143644&version=1.1}

\url{https://github.com/aaronrkaufman/compactness/tree/e91519e15fd3ca55f57fc21323bb12ed5a66305b/data}

\url{https://alarm-redist.org/redistmetrics/}
}
\usage{
compact_train
}
\description{
Clean data used to model the DESPLIM compactness metric, based on
the data provided by Kaufman et al. (2021). The metrics have been computed
using the \code{redistmetrics} package.
}
\keyword{datasets}
