% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desplim_compactness.R
\name{desplim_compactness}
\alias{desplim_compactness}
\title{Calculate the 'desplim' compactness of an \code{sf} POLYGON object}
\usage{
desplim_compactness(input, keep_metrics = FALSE)
}
\arguments{
\item{input}{object of class \code{sf} of type POLYGON or MULTIPOLYGON.}

\item{keep_metrics}{logical; should all the compactness metrics be kept?
Default is \code{FALSE}.}
}
\value{
An sf object of type POLYGON matching \code{input}, with the compactness
metric(s) added.
}
\description{
Based on the approach described in Kaufmann et al. (2021), this
function computes the compactness of an \code{sf} POLYGON object.
}
\details{
Please see the dedicated vignette for information about the
compactness metric. If \code{input} contains geometries of type MULTIPOLYGON, they
are cast to POLYGON before computing.
}
\examples{
# Create simple plygon
polygon_coords <- list(matrix(
 c(0, 0, 0, 1, 1, 3, -1, 2, -3, 1, -2, 0, 0, 0),
 ncol = 2,
 byrow = TRUE
))
simple_polygon <- sf::st_sfc(sf::st_polygon(polygon_coords), crs = 4326)
input_sf <- sf::st_sf(geometry = simple_polygon)
# Calculate compactness
compactness_result <- desplim_compactness(input_sf, keep_metrics = TRUE)
plot(input_sf)
print(sf::st_drop_geometry(compactness_result))
}
