% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desplim_merge.R
\name{desplim_merge}
\alias{desplim_merge}
\title{Deterministically merge a set of polygons with possible parameters defined by
the user.}
\usage{
desplim_merge(
  input_polygons,
  input_buildings = NULL,
  compact_method = "desplim",
  compact_threshold = 0.7,
  compact_allow = 0.2,
  area_threshold = 10000,
  enclosed_threshold = 0.8,
  building_threshold = 5
)
}
\arguments{
\item{input_polygons}{object of class sf of type POLYGON to be merged.}

\item{input_buildings}{object of class sf of type POLYGON or MULTIPOLYGON.
An optional sf object representing buildings which should be considered when
merging. Default is \code{NULL}.}

\item{compact_method}{string; chosen method for calculating compactness of
polygons. Default is \code{"desplim"}, using the internal \code{desplim_compactness}
function. Other options include \code{"polsby"}, \code{"schwartz"} and \code{"convex_hull"}
from the \code{redistmetrics} package.}

\item{compact_threshold}{numerical; the baseline expected compactness of the
output polygons. Default is 0.7.}

\item{compact_allow}{numerical; the fraction of polygons allowed to fall
below \code{compact_tolerance}. Default is 0.2.}

\item{area_threshold}{numerical; the minimum required area of merged
polygons. Default is 10000 units.}

\item{enclosed_threshold}{numerical; the fraction of a polygon border shared
with a single neighbour for it to be considered enclosed (and flagged for
merging). Default is 0.8.}

\item{building_threshold}{integer; the minimum number of required buildings
within merged polygons if \code{input_buildings} is provided. Default is 5, but
sis set to 0 if \code{input_buildings} is not provided.}
}
\value{
An sf object of POLYGONs resulting from the deterministic merge.
}
\description{
Function for merging a set of polygons, with optinal rules
defined by the user. The merging process is deterministic, leading to
reproducible results on subsequent runs with the same input data and rules.
}
\details{
There are no requirements for all polygons being adjacent to each
other. Although named \code{input_buildings}, any sf object of type POLYGON or
MULTIPOLYGON can be provided for this argument. For details about the
compactness metric, please see the dedicated vignette. It is not recommended
to set \code{compact_allow} too close to 1 and \code{compact_tolerance} too low.
This is especially true if \code{compact_method} is set to \code{"polsby"}.
}
