% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desplim_split_merge.r
\name{desplim_split_merge}
\alias{desplim_split_merge}
\title{Apply the DESPLIM algorithm to a set of input polygons and lines.}
\usage{
desplim_split_merge(
  input_polygon,
  input_lines,
  input_buildings = NULL,
  line_type_identifier = NULL,
  line_type_hierarchy = NULL,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{input_polygon}{object of class sf of type POLYGON to be merged.}

\item{input_lines}{object of class sf of type LINESTRING or MULTILINESTRING
to be used for splitting.}

\item{input_buildings}{object of class sf of type POLYGON or MULTIPOLYGON.
An optional sf object representing buildings which should be considered when
splitting and merging. Default is \code{NULL}.}

\item{line_type_identifier}{string; the name of the column in \code{input_lines}
which contains the line type. Default is \code{NULL}.}

\item{line_type_hierarchy}{list; hierarchical structure of line types used
for splitting. Default is \code{NULL}. All types can be accessed with \code{"all"}.}

\item{parallel}{logical; whether to run the split and merging steps in
parallel. If \code{TRUE}, uses the \code{future} package, where the number of workers
should be set using the \code{plan} argument. Default is \code{FALSE}.}

\item{...}{additional arguments passed to \code{desplim_split}, \code{desplim_merge}
and \code{desplim_connect_border}.}
}
\value{
An sf object of POLYGONs resulting from the merge.
}
\description{
Apply the DESPLIM algorithm to a set of input polygons and lines.
}
\details{
The function applies the DESPLIM algorithm to a set of input
polygons and lines. The splitting and merging steps are performed in line
with the \code{desplim_split} and \code{desplim_merge} functions for each iteration as
defined by \code{line_type_hierarchy}.
}
